Add transactionality in read database operations. Refactor services
