+++ src/stations/services/database/transactions/database-transaction.service.ts
@@ -1,0 +1,24 @@
+import { Injectable } from '@nestjs/common';
+import { InjectConnection } from '@nestjs/mongoose';
+import { Connection } from 'mongoose';
+
+@Injectable()
+export class DatabaseTransactionService {
+  constructor(@InjectConnection() private readonly connection: Connection) {}
+
+  async execute<T>(handler: (session: any) => Promise<T>): Promise<T> {
+    const session = await this.connection.startSession();
+    session.startTransaction();
+
+    try {
+      const result = await handler(session);
+      await session.commitTransaction();
+      return result;
+    } catch (error) {
+      await session.abortTransaction();
+      throw error;
+    } finally {
+      session.endSession();
+    }
+  }
+}
